requirements_test.txt
pyairbnk @ https://github.com/Moballo-LLC/pyairbnk/archive/refs/tags/v1.0.0.tar.gz
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component
aiousbwatcher
pyserial
ruff
mypy
